Update README.md
This is all you need to deploy your very own copy of the genome browser.  These instructions are edits and improvements on Meng's work which you can find here.

https://github.com/icebert/docker_ucsc_genome_browser

Note that I am currently a Dphil student at the University of Oxford and as such I am doing this for academic research and testing.  See above for acquiring a commercial license.